requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
dill
Pillow
streamlit
opencv-python
keras
tensorflow
streamlit-webrtc
#-e .